\name{lm_t_pc}
\alias{lm_t_pc}
\title{
Comparison between correlations of two groups
}
\description{Comparison between correlations of two groups}
\usage{lm_t_pc(data1, data2, label1, label2)}
\arguments{
  \item{data1}{The result generated from find.side.gene()}
  \item{data2}{The result generated from find.side.gene()}
  \item{lable1}{The character your want to define data1}
  \item{lable1}{The character your want to define data1}
  }
\details{lm_t_pc illustrate two linear relations on one graph}
\value{
datas: The result generated from find.side.gene();
labels: characters, like, "test1"
}
\author{
Songchen Yang
}
\examples{
\dontrun{
lm_t_pc(data1, data2, test1, test2)
}
}

